refactored gene routes
