+++ nsweb/controllers/genes.py
@@ -1,11 +1,9 @@
-from flask import (Blueprint, render_template, abort, send_file)
+from flask import Blueprint, render_template, url_for
 from nsweb.models.genes import Gene
-from nsweb.tasks import make_scatterplot
-from nsweb.initializers import settings
 from nsweb.controllers import error_page
-from nsweb.controllers.decode import decode_analysis_image
+from nsweb.api.decode import decode_analysis_image
 import json
-from os.path import join, basename, exists
+
 
 bp = Blueprint('genes', __name__, url_prefix='/genes')
 
@@ -23,7 +21,7 @@ def show(symbol):
     image = gene.images[0]
     # Run decoder if it hasn't been run before
     dec = decode_analysis_image(image.id)
-    url = '/images/%s' % image.id
+    url = url_for('api_images.download', val=image.id)
     images = [{
         'id': image.id,
         'name': symbol,
@@ -34,21 +32,3 @@ def show(symbol):
     }]
     return render_template('genes/show.html', gene=gene,
                            images=json.dumps(images), image_id=dec.uuid)
-
-
-@bp.route('/<string:val>/scatter/<string:analysis>.png')
-def get_scatter(val, analysis):
-    outfile = join(
-        settings.DECODING_SCATTERPLOTS_DIR, val + '_' + analysis + '.png')
-    if not exists(outfile):
-        """ Return .png of scatter plot between the uploaded image and
-        specified analysis. """
-        gene = Gene.query.filter_by(symbol=val).first()
-        if gene is None:
-            abort(404)
-        make_scatterplot.delay(
-            gene.images[0].image_file, analysis, gene.symbol,
-            x_lab='%s expression level' % gene.symbol, outfile=outfile,
-            gene_masks=True).wait()
-    return send_file(outfile, as_attachment=False,
-                     attachment_filename=basename(outfile))
